Finalizing string to filter methods
