+++ app/src/context/SearchProvider.tsx
@@ -5,33 +5,15 @@ import {
   DEFAULT_SEARCH_SORT,
   DEFAULT_SEARCH_TERM,
 } from "../config/constants";
-import { GeneralSearchManager, SearchManager } from "../utils/searchManager";
-import { Filter } from "../types/FilterType";
+import {
+  GeneralSearchManager,
+  SearchManager,
+  stringToFilter,
+} from "../utils/searchManager";
 
 interface SearchContextType {
   searchManager: SearchManager;
 }
-
-export const stringToFilter = (filtersString: string | null): Filter[] => {
-  if (!filtersString) return [];
-
-  return filtersString
-    .split(",")
-    .map((filterPair) => {
-      const match = filterPair.match(/^\[(\w+)=(.+)\]$/);
-      if (match) {
-        const [, filter, value] = match;
-        return { filter, value };
-      }
-      return null;
-    })
-    .filter((filter): filter is Filter => filter !== null);
-};
-
-export const filterToString = (filters: Filter[]): string => {
-  if (!filters || filters.length === 0) return "";
-  return filters.map(({ filter, value }) => `[${filter}=${value}]`).join("");
-};
 
 const SearchContext = createContext<SearchContextType | undefined>(undefined);
